refactor(provisioner): split gdc_vmseries_ngfw into common/assets modules (#878)
Clear the S104 (>500 lines) smell on gdc_vmseries_ngfw.py (828 -> 375 lines) via
a cohesive three-module split: gdc_vmseries_common (label/product constants,
the _VMSeriesNames projection, deterministic resource-name helpers,
contextlib_suppress) and gdc_vmseries_assets (Secret-Manager + bootstrap-ISO +
VM/disk/interface manifest builders), leaving lifecycle orchestration
(apply_ngfw/destroy_ngfw/run_power_operation) in gdc_vmseries_ngfw. The import
graph is common <- assets <- ngfw (no cycle); ngfw re-imports the helpers its
orchestration calls by bare name, preserving the patch("gdc_vmseries_ngfw.X")
test seams. Full provisioner suite: 754 passed, 8 skipped.
